🏗️ Reorganisation des specs de raccordement (#3046)
