+++ packages/specifications/src/raccordement/dateDeMiseEnService/stepDefinitions/dateMiseEnService.given.ts
@@ -4,7 +4,7 @@ import { mediator } from 'mediateur';
 import { DateTime } from '@potentiel-domain/common';
 import { Raccordement } from '@potentiel-domain/laureat';
 
-import { PotentielWorld } from '../../potentiel.world';
+import { PotentielWorld } from '../../../potentiel.world';
 
 EtantDonné(
   'une date de mise en service pour le dossier de raccordement du projet lauréat',
