+++ packages/specifications/src/raccordement/demandeComplèteDeRaccordement/fixtures/transmettreDemandeComplèteDeRaccordement.fixture.ts
@@ -4,20 +4,20 @@ import { DocumentProjet } from '@potentiel-domain/document';
 import { Raccordement } from '@potentiel-domain/laureat';
 import { DateTime } from '@potentiel-domain/common';
 
-import { AbstractFixture } from '../../fixture';
-import { convertStringToReadableStream } from '../../helpers/convertStringToReadable';
+import { AbstractFixture } from '../../../fixture';
+import { convertStringToReadableStream } from '../../../helpers/convertStringToReadable';
 
 type PièceJustificative = { format: string; content: ReadableStream };
 
-interface ModifierDemandeComplèteRaccordement {
+interface TransmettreDemandeComplèteRaccordement {
   dateQualification: string;
   référenceDossier: string;
   accuséRéception: PièceJustificative;
 }
 
-export class ModifierDemandeComplèteRaccordementFixture
-  extends AbstractFixture<ModifierDemandeComplèteRaccordement>
-  implements ModifierDemandeComplèteRaccordement
+export class TransmettreDemandeComplèteRaccordementFixture
+  extends AbstractFixture<TransmettreDemandeComplèteRaccordement>
+  implements TransmettreDemandeComplèteRaccordement
 {
   #dateQualification!: string;
   get dateQualification(): string {
@@ -44,10 +44,10 @@ export class ModifierDemandeComplèteRaccordementFixture
   }
 
   créer(
-    partialFixture: Partial<Readonly<ModifierDemandeComplèteRaccordement>> & {
+    partialFixture: Partial<Readonly<TransmettreDemandeComplèteRaccordement>> & {
       identifiantProjet: string;
     },
-  ): Readonly<ModifierDemandeComplèteRaccordement> {
+  ): Readonly<TransmettreDemandeComplèteRaccordement> {
     const content = faker.word.words();
     const fixture = {
       dateQualification: faker.date.recent().toISOString(),
@@ -66,6 +66,19 @@ export class ModifierDemandeComplèteRaccordementFixture
     this.#identifiantProjet = fixture.identifiantProjet;
     this.aÉtéCréé = true;
     return fixture;
+  }
+
+  mapExempleToFixtureValues(exemple: Record<string, string>) {
+    const values: Partial<TransmettreDemandeComplèteRaccordement> = {};
+    const dateQualification = exemple['La date de qualification'];
+    const référenceDossier = exemple['La référence du dossier de raccordement'];
+    if (dateQualification) {
+      values.dateQualification = new Date(dateQualification).toISOString();
+    }
+    if (référenceDossier) {
+      values.référenceDossier = référenceDossier;
+    }
+    return values;
   }
 
   mapToExpected(référenceDossier?: string) {
