+++ packages/specifications/src/raccordement/dossierRaccordement/fixtures/modifierRéférenceDossierRaccordement.fixture.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { AbstractFixture } from '../../fixture';
+import { AbstractFixture } from '../../../fixture';
 
 interface ModifierRéférenceDossierRaccordement {
   référenceDossier: string;
