+++ packages/specifications/src/raccordement/dossierRaccordement/supprimerDossierDuRaccordement.feature
@@ -1,4 +1,5 @@
 # language: fr
+@select
 Fonctionnalité: Supprimer un dossier du raccordement d'un projet
 
     Contexte:
