+++ packages/specifications/src/raccordement/gestionnaireRéseau/stepDefinitions/gestionnaireRéseauRaccordement.when.ts
@@ -2,11 +2,10 @@ import { When as Quand } from '@cucumber/cucumber';
 import { mediator } from 'mediateur';
 
 import { GestionnaireRéseau } from '@potentiel-domain/reseau';
-import { Raccordement } from '@potentiel-domain/laureat';
 import { Role } from '@potentiel-domain/utilisateur';
+import { Raccordement } from '@potentiel-domain/laureat';
 
-import { PotentielWorld } from '../../potentiel.world';
-
+import { PotentielWorld } from '../../../potentiel.world';
 Quand(
   `le porteur modifie le gestionnaire de réseau du projet avec un gestionnaire non référencé`,
   async function (this: PotentielWorld) {
@@ -48,45 +47,6 @@ Quand(
       raisonSocialGestionnaireRéseau,
     );
     await modifierGestionnaireRéseauRaccordement.call(this, codeEIC, Role.dreal);
-  },
-);
-
-Quand(
-  `le porteur supprime le dossier de raccordement pour le projet lauréat`,
-  async function (this: PotentielWorld) {
-    const { identifiantProjet } = this.lauréatWorld;
-    const { référenceDossier } = this.raccordementWorld;
-
-    try {
-      await mediator.send<Raccordement.SupprimerDossierDuRaccordementUseCase>({
-        type: 'Lauréat.Raccordement.UseCase.SupprimerDossierDuRaccordement',
-        data: {
-          identifiantProjetValue: identifiantProjet.formatter(),
-          référenceDossierValue: référenceDossier,
-        },
-      });
-    } catch (e) {
-      this.error = e as Error;
-    }
-  },
-);
-
-Quand(
-  `le porteur supprime le dossier de raccordement pour le projet lauréat avec pour référence {string}`,
-  async function (this: PotentielWorld, référenceDossier) {
-    const { identifiantProjet } = this.lauréatWorld;
-
-    try {
-      await mediator.send<Raccordement.SupprimerDossierDuRaccordementUseCase>({
-        type: 'Lauréat.Raccordement.UseCase.SupprimerDossierDuRaccordement',
-        data: {
-          identifiantProjetValue: identifiantProjet.formatter(),
-          référenceDossierValue: référenceDossier,
-        },
-      });
-    } catch (e) {
-      this.error = e as Error;
-    }
   },
 );
