+++ packages/specifications/src/raccordement/propositionTechniqueEtFinancière/fixtures/modifierPropositionTechniqueEtFinancière.fixture.ts
@@ -4,8 +4,8 @@ import { DateTime } from '@potentiel-domain/common';
 import { DocumentProjet } from '@potentiel-domain/document';
 import { Raccordement } from '@potentiel-domain/laureat';
 
-import { AbstractFixture } from '../../fixture';
-import { convertStringToReadableStream } from '../../helpers/convertStringToReadable';
+import { AbstractFixture } from '../../../fixture';
+import { convertStringToReadableStream } from '../../../helpers/convertStringToReadable';
 
 type PièceJustificative = { format: string; content: ReadableStream };
