+++ packages/specifications/src/raccordement/propositionTechniqueEtFinancière/fixtures/transmettrePropositionTechniqueEtFinancière.fixture.ts
@@ -4,8 +4,8 @@ import { DateTime } from '@potentiel-domain/common';
 import { DocumentProjet } from '@potentiel-domain/document';
 import { Raccordement } from '@potentiel-domain/laureat';
 
-import { convertStringToReadableStream } from '../../helpers/convertStringToReadable';
-import { AbstractFixture } from '../../fixture';
+import { convertStringToReadableStream } from '../../../helpers/convertStringToReadable';
+import { AbstractFixture } from '../../../fixture';
 
 type PièceJustificative = { format: string; content: ReadableStream };
