+++ packages/specifications/src/raccordement/propositionTechniqueEtFinancière/propositionTechniqueEtFinancière.world.ts
