+++ packages/specifications/src/raccordement/propositionTechniqueEtFinancière/stepDefinitions/propositionTechniqueEtFinancière.when.ts
@@ -4,7 +4,7 @@ import { mediator } from 'mediateur';
 import { Raccordement } from '@potentiel-domain/laureat';
 import { IdentifiantProjet } from '@potentiel-domain/common';
 
-import { PotentielWorld } from '../../potentiel.world';
+import { PotentielWorld } from '../../../potentiel.world';
 
 Quand(
   `le porteur transmet une proposition technique et financière pour le dossier de raccordement du projet lauréat`,
