+++ packages/specifications/src/raccordement/supprimerRaccordement.feature
@@ -1,4 +1,5 @@
 # language: fr
+@select
 Fonctionnalité: Supprimer le raccordement d'un projet
 
     Contexte:
